Further cleanup, move files around.
- Make struct fields private and have getters instead.
