+++ src/image_helper.rs
@@ -1,5 +1,5 @@
 pub mod ascii_image_options;
-pub mod image_converter;
+pub mod image_converters;
 pub mod image_data;
 pub mod image_error;
 pub mod image_writer;
